requirements.txt
numpy==1.24.4
pandas==2.2.3
scikit-learn==1.5.2

torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1

transformers==4.46.3
tokenizers==0.20.3

opencv-python==4.10.0.84
Pillow==11.0.0

librosa==0.10.2.post1
soundfile==0.12.1

lime==0.2.0.1
shap==0.46.0

matplotlib==3.9.2
seaborn==0.13.2

pyyaml==6.0.2
tqdm==4.66.6
python-dotenv==1.0.1
joblib==1.4.2
loguru==0.7.2

requests==2.32.3
praw==7.8.1
openpyxl==3.1.5

